+++ app/javascript/components/Tasks.tsx
@@ -1,9 +1,13 @@
 import { useState } from "react";
 import { Link } from "react-router-dom";
-import { useTasksQuery, TaskStateEnum } from "../graphql/generated";
+import { useTasksQuery } from "../graphql/generated";
 import { TaskStateLabel, TaskState } from "./Enum";
+import { useLimitOnDisplay } from "../hooks/useLimitOnDisplay";
+import { useLimitMessage } from "../hooks/useLimitMessage";
 
 const Tasks = () => {
+  const { limitOnDisplay } = useLimitOnDisplay();
+  const { limitMessage } = useLimitMessage();
   const [searchTitle, setSearchTitle] = useState("");
   const [searchState, setSearchState] = useState("");
   const [title, setTitle] = useState("");
@@ -32,45 +36,6 @@ const Tasks = () => {
   };
 
   const resetTaskSearch = () => setTitle("");
-
-  const formatDate = (date: Date): string => {
-    const y: number = date.getFullYear();
-    const m: string = ("00" + (date.getMonth() + 1)).slice(-2);
-    const d: string = ("00" + date.getDate()).slice(-2);
-    return `${y + "-" + m + "-" + d}`;
-  };
-
-  const diffLimitOn = (limitOn: any, state: TaskStateEnum | undefined) => {
-    if (limitOn == null) return;
-    if (state == "finished") return;
-    const limitOnParse: Date = new Date(limitOn);
-    const today: Date = new Date(formatDate(new Date()));
-    const diffDay: number = Math.floor(
-      (limitOnParse.getTime() - today.getTime()) / 86400000
-    );
-    const limitMessage = () => {
-      if (diffDay > 3 || diffDay == undefined) {
-        return;
-      } else if (diffDay < 1 && diffDay > -1) {
-        return "-期限当日です";
-      } else if (diffDay <= 3 && diffDay > 0) {
-        return `-期限${diffDay}日前です`;
-      } else {
-        return "-期限が過ぎています";
-      }
-    };
-
-    return limitMessage();
-  };
-
-  const limitOnDisplay = (limitOn: any, state: TaskStateEnum | undefined) => {
-    if (state == "finished") return;
-    if (limitOn) {
-      return `-${limitOn}`;
-    } else {
-      return "-期限未登録";
-    }
-  };
 
   return (
     <div>
@@ -109,7 +74,7 @@ const Tasks = () => {
             <Link to={`/tasks/${task?.node?.id}`}>
               {TaskStateLabel(task?.node?.state)}-{task?.node?.title}
               {limitOnDisplay(task?.node?.limitOn, task?.node?.state)}
-              {diffLimitOn(task?.node?.limitOn, task?.node?.state)}
+              {limitMessage(task?.node?.limitOn, task?.node?.state)}
             </Link>
           </li>
         ))}
